+++ Sources/PulseUI/Features/Console/List/ConsoleListContentView.swift
@@ -2,6 +2,8 @@
 //
 // Copyright (c) 2020–2023 Alexander Grebenyuk (github.com/kean).
 
+#if !PULSE_STANDALONE_APP
+
 import CoreData
 import Pulse
 import Combine
@@ -155,4 +157,6 @@ struct ConsoleStaticList: View {
 #endif
     }
 }
+#endif
+
 #endif